💚 Fix tests
